devops-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
openai==1.3.5
python-dotenv==1.0.0
httpx==0.25.2
python-jose==3.3.0
Pillow==10.1.0
aiofiles==23.2.1
